requirements.txt
aioitertools==0.4.0
aiohttp==3.5.4
beautifulsoup4==4.7.1
dataclasses==0.6
multidict==4.5.2
typing==3.6.6
yarl==1.3.0
aiosqlite==0.10.0
tortoise-orm==0.11.12
PuLP==1.6.10
sympy==1.4
